% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{psa001}
\alias{psa001}
\title{PSA001}
\usage{
psa001(openRmd = FALSE)
}
\arguments{
\item{openRmd}{Whether to save an analysis Rmd file to your working directory and open it.}
}
\description{
Load all files for PSA001 Social Faces
}
