% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa001_agg.R
\docType{data}
\name{psa001_agg}
\alias{psa001_agg}
\title{PSA001 Aggregated Data}
\format{
A data frame with 1320 rows and 15 variables:
\describe{
}
}
\source{
\url{https://doi.org/10.1038/s41562-020-01007-2}
}
\usage{
psa001_agg
}
\description{
Codebook for PSA001 Aggregated Data (Psych-DS 0.1.0)
}
\details{
Dataset Parameters
\itemize{
\item name: PSA001 Aggregated Data
\item schemaVersion: Psych-DS 0.1.0
\item author: PSA
\item description: Average ratings for each of 120 stimuli for each of 11 regions.
\item license: CC-BY 4.0
}

Column Parameters
\itemize{
\item region (string)
\itemize{
\item Levels
\itemize{
\item Africa
\item Asia
\item Australia & New Zealand
\item Central America & Mexico
\item Eastern Europe
\item Middle East
\item Scandinavia
\item South America
\item UK
\item USA & Canada
\item Western Europe
}
\item Ordered: FALSE
}
\item stim_id (string)
\item aggressive (float)
\item attractive (float)
\item caring (float)
\item confident (float)
\item dominant (float)
\item emostable (float)
\item intelligent (float)
\item mean (float)
\item responsible (float)
\item sociable (float)
\item trustworthy (float)
\item unhappy (float)
\item weird (float)
}
}
\keyword{datasets}
