% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa001_cfd_faces.R
\docType{data}
\name{psa001_cfd_faces}
\alias{psa001_cfd_faces}
\title{PSA001 Stimulus Demographics}
\format{
A data frame with 120 rows and 4 variables:
\describe{
}
}
\source{
\url{CFD}
}
\usage{
psa001_cfd_faces
}
\description{
Codebook for PSA001 Stimulus Demographics (Psych-DS 0.1.0)
}
\details{
Dataset Parameters
\itemize{
\item name: PSA001 Stimulus Demographics
\item schemaVersion: Psych-DS 0.1.0
\item NA: {empty}
\item author: CFD
\item description: Demographic characteristics for the 120 faces from the Chicago Face Database.
\item license: CC-BY 4.0
}

Column Parameters
\itemize{
\item Target (string): Stimulus ID
\item Race (string): Racial category
\itemize{
\item Levels
\itemize{
\item A: Asian
\item B: Black
\item L: Latino
\item W: White
}
\item Ordered: FALSE
}
\item Gender (string)
\itemize{
\item Levels
\itemize{
\item M: Male
\item F: Female
}
\item Ordered: FALSE
}
\item Age (float): Age in years
}
}
\keyword{datasets}
